feat(styles): update token with the latest Horizon design and split tokenizer doc [ci visual]
