+++ packages/styles/src/theming/common/token/_sap_fiori.scss
@@ -1,0 +1,17 @@
+:root {
+  --fdToken_Text_Shadow: none;
+  --fdToken_Border_Radius: 0.25rem;
+  --fdToken_Text_Color: var(--sapTextColor);
+  --fdToken_Background_Color_Hover: var(--sapButton_Hover_Background);
+  --fdToken_Border_Color_Hover: var(--sapButton_Hover_BorderColor);
+  --fdToken_Outline_Offset: -0.1875rem;
+  --fdToken_Outline_Offset_Compact: -0.1875rem;
+  --fdToken_Background_Color_Read_Only: var(--sapButton_TokenBackground);
+  --fdToken_Border_Color_Read_Only: var(--sapButton_TokenBorderColor);
+  --fdToken_Border_Color_Selected_Focus: var(--sapContent_ContrastFocusColor);
+  --fdToken_Selected_Border: var(--sapButton_BorderWidth) solid var(--sapButton_Selected_BorderColor);
+  --fdToken_Selected_Font_Family: var(--sapFontFamily);
+  --fdToken_Selected_Border_BottomStyle: var(--sapButton_Selected_Background);
+  --fdToken_ReadOnly_Padding_Inline: 0.375rem;
+  --fdToken_Focus_Border_Radius: 0.25rem;
+}
