+++ packages/styles/src/theming/common/token/_sap_fiori_hc.scss
@@ -1,0 +1,7 @@
+@import "./sap_fiori";
+
+:root {
+  --fdToken_Text_Shadow: 0 0 0.125rem #000;
+  --fdToken_Background_Color_Read_Only: var(--sapField_ReadOnly_Background);
+  --fdToken_Border_Color_Read_Only: var(--sapField_ReadOnly_BorderColor);
+}
