+++ packages/styles/src/theming/sap_fiori_3.scss
@@ -35,6 +35,7 @@
 @import "./common/avatar/sap_fiori";
 @import "./common/avatar-group/sap_fiori";
 @import "./common/message-popover/sap_fiori";
+@import "./common/token/sap_fiori";
 @import "./sap_fiori_3/css_variables";
 
 :root {
@@ -49,21 +50,6 @@
   --fdDynamicPage_Button_Group_Background: var(--sapHighlightColor);
   --fdDynamicPage_Content_List_Background: var(--sapGroup_ContentBackground);
   --fdDynamicPage_Content_Transparent_Background: transparent;
-
-  /* Token */
-  --fdToken_Text_Shadow: none;
-  --fdToken_Border_Radius: 0.25rem;
-  --fdToken_Text_Color: var(--sapTextColor);
-  --fdToken_Padding_Left: 0.3125rem;
-  --fdToken_Background_Color_Hover: var(--sapButton_Hover_Background);
-  --fdToken_Border_Color_Hover: var(--sapButton_Hover_BorderColor);
-  --fdToken_Box_Shadow_Hover: none;
-  --fdToken_Box_Shadow_Selected_Hover: none;
-  --fdToken_Outline_Offset: -0.1875rem;
-  --fdToken_Outline_Offset_Compact: -0.1875rem;
-  --fdToken_Background_Color_Read_Only: var(--sapButton_TokenBackground);
-  --fdToken_Border_Color_Read_Only: var(--sapButton_TokenBorderColor);
-  --fdToken_Border_Color_Selected_Focus: var(--sapContent_ContrastFocusColor);
 
   /* TODO: DEPRECATED Side-Nav */
   --fdSideNav_Text_Shadow: none;
