+++ packages/styles/src/theming/sap_horizon.scss
@@ -37,6 +37,7 @@
 @import "./common/avatar/sap_horizon";
 @import "./common/avatar-group/sap_horizon";
 @import "./common/message-popover/sap_horizon";
+@import "./common/token/sap_horizon";
 @import "./sap_horizon/css_variables";
 
 :root {
@@ -58,21 +59,6 @@
   --fdDynamicPage_Button_Group_Background: var(--sapObjectHeader_BorderColor);
   --fdDynamicPage_Content_List_Background: var(--sapGroup_ContentBackground);
   --fdDynamicPage_Content_Transparent_Background: transparent;
-
-  /* Token */
-  --fdToken_Text_Shadow: none;
-  --fdToken_Border_Radius: var(--sapButton_BorderCornerRadius);
-  --fdToken_Text_Color: var(--sapButtonTextColor);
-  --fdToken_Padding_Left: 0.25rem;
-  --fdToken_Background_Color_Hover: var(--sapButton_TokenBackground);
-  --fdToken_Border_Color_Hover: var(--sapButton_TokenBorderColor);
-  --fdToken_Box_Shadow_Hover: var(--sapContent_Interaction_Shadow);
-  --fdToken_Box_Shadow_Selected_Hover: var(--sapContent_Interaction_Shadow);
-  --fdToken_Outline_Offset: 0.0625rem;
-  --fdToken_Outline_Offset_Compact: 0;
-  --fdToken_Background_Color_Read_Only: var(--sapButton_TokenBackground);
-  --fdToken_Border_Color_Read_Only: var(--sapButton_TokenBorderColor);
-  --fdToken_Border_Color_Selected_Focus: var(var(--sapContent_FocusColor));
 
   /* TODO: DEPRECATED Side-Nav */
   --fdSideNav_Text_Shadow: none;
