+++ packages/styles/src/theming/sap_horizon_hcb.scss
@@ -37,6 +37,7 @@
 @import "./common/avatar/sap_horizon_hc";
 @import "./common/avatar-group/sap_horizon";
 @import "./common/message-popover/sap_horizon_hc";
+@import "./common/token/sap_horizon_hc";
 @import "./sap_horizon_hcb/css_variables";
 
 :root {
@@ -51,21 +52,6 @@
   --fdDynamicPage_Button_Group_Background: var(--sapObjectHeader_BorderColor);
   --fdDynamicPage_Content_List_Background: var(--sapBackgroundColor);
   --fdDynamicPage_Content_Transparent_Background: var(--sapBackgroundColor);
-
-  /* Token */
-  --fdToken_Text_Shadow: none;
-  --fdToken_Border_Radius: var(--sapButton_BorderCornerRadius);
-  --fdToken_Text_Color: var(--sapButtonTextColor);
-  --fdToken_Padding_Left: 0.25rem;
-  --fdToken_Background_Color_Hover: var(--sapButton_Hover_Background);
-  --fdToken_Border_Color_Hover: var(--sapButton_TokenBorderColor);
-  --fdToken_Box_Shadow_Hover: var(--sapContent_Interaction_Shadow);
-  --fdToken_Box_Shadow_Selected_Hover: var(--sapContent_Interaction_Shadow);
-  --fdToken_Outline_Offset: -0.1875rem;
-  --fdToken_Outline_Offset_Compact: -0.1875rem;
-  --fdToken_Background_Color_Read_Only: var(--sapField_ReadOnly_Background);
-  --fdToken_Border_Color_Read_Only: var(--sapField_ReadOnly_BorderColor);
-  --fdToken_Border_Color_Selected_Focus: var(--sapContent_ContrastFocusColor);
 
   /* TODO: DEPRECATED Side-Nav */
   --fdSideNav_Text_Shadow: none;
