+++ packages/styles/stories/Components/token/primary.example.html
@@ -1,24 +1,80 @@
-<span class="fd-token" role="button" tabindex="0">
-    <span class="fd-token__text">
-        Bibendum
-    </span>
-    <button class="fd-token__close" tabindex="-1" aria-label="Bibendum token close"></button>
-</span>
-<span class="fd-token" role="button" tabindex="0">
-    <span class="fd-token__text">
-        Lorem
-    </span>
-    <button class="fd-token__close" tabindex="-1" aria-label="Lorem token close"></button>
-</span>
-<span class="fd-token" role="button" tabindex="0">
-    <span class="fd-token__text">
-        Dolor
-    </span>
-    <button class="fd-token__close" tabindex="-1" aria-label="Dolor token close"></button>
-</span>
-<span class="fd-token" role="button" tabindex="0">
-    <span class="fd-token__text">
-        Filter
-    </span>
-    <button class="fd-token__close" tabindex="-1" aria-label="Filter token close"></button>
-</span>
+<style>
+    .docs-tokens {
+        width: 100%;
+    }
+    .docs_tokens_item_wrapper {
+        display: flex;
+    }
+    .docs-tokens-container {
+        display: grid;
+        grid-template-columns: repeat(auto-fit, minmax(80px, 1fr));
+        column-gap: 0.1rem;
+        row-gap: 0.1rem;
+        padding: 1rem;
+        justify-items: right;
+        align-items: center;
+    }
+</style>
+
+
+<section class="docs-tokens">
+    <div class="docs-tokens-container">
+        <div></div>
+        <div>Regular</div>
+        <div>Selected</div>
+    </div>
+    <div class="docs-tokens-container">
+        <div>Normal</div>
+        <span class="fd-token" role="button" tabindex="0">
+            <span class="fd-token__text">tokenopolis</span>
+            <button class="fd-token__close" tabindex="-1" aria-label="Bibendum token close"></button>
+        </span>
+        <span class="fd-token fd-token--selected" role="button" tabindex="0">
+            <span class="fd-token__text">tokenopolis</span>
+            <button class="fd-token__close" tabindex="-1" aria-label="Bibendum token close"></button>
+        </span>
+    </div>
+    <div class="docs-tokens-container">
+        <div>Hover</div>
+        <span class="fd-token is-hover" role="button" tabindex="0">
+            <span class="fd-token__text">tokenopolis</span>
+            <button class="fd-token__close" tabindex="-1" aria-label="Bibendum token close"></button>
+        </span>
+        <span class="fd-token fd-token--selected is-hover" role="button" tabindex="0">
+            <span class="fd-token__text">tokenopolis</span>
+            <button class="fd-token__close" tabindex="-1" aria-label="Bibendum token close"></button>
+        </span>
+    </div>
+    <div class="docs-tokens-container">
+        <div>Focus</div>
+        <span class="fd-token is-focus" role="button" tabindex="0">
+            <span class="fd-token__text">tokenopolis</span>
+            <button class="fd-token__close" tabindex="-1" aria-label="Bibendum token close"></button>
+        </span>
+        <span class="fd-token fd-token--selected is-focus" role="button" tabindex="0">
+            <span class="fd-token__text">tokenopolis</span>
+            <button class="fd-token__close" tabindex="-1" aria-label="Bibendum token close"></button>
+        </span>
+    </div>
+    <div class="docs-tokens-container">
+        <div>Hover&Focus</div>
+        <span class="fd-token is-hover is-focus" role="button" tabindex="0">
+            <span class="fd-token__text">tokenopolis</span>
+            <button class="fd-token__close" tabindex="-1" aria-label="Bibendum token close"></button>
+        </span>
+        <span class="fd-token fd-token--selected is-hover is-focus" role="button" tabindex="0">
+            <span class="fd-token__text">tokenopolis</span>
+            <button class="fd-token__close" tabindex="-1" aria-label="Bibendum token close"></button>
+        </span>
+    </div>
+    <div class="docs-tokens-container">
+        <div>Read-only</div>
+        <span class="fd-token fd-token--readonly" role="button" tabindex="0">
+            <span class="fd-token__text">tokenopolis</span>
+        </span>
+        <span></span>
+        <!-- <span class="fd-token fd-token--selected fd-token--readonly" role="button" tabindex="0">
+            <span class="fd-token__text">tokenopolis</span>
+        </span> -->
+    </div>
+</section>
