--- packages/styles/stories/Components/token/read-only-token.example.html
@@ -1,20 +1,0 @@
-<span class="fd-token fd-token--readonly" role="button" tabindex="0">
-    <span class="fd-token__text">
-        Bibendum
-    </span>
-</span>
-<span class="fd-token fd-token--readonly" role="button" tabindex="0">
-    <span class="fd-token__text">
-        Lorem
-    </span>
-</span>
-<span class="fd-token fd-token--readonly" role="button" tabindex="0">
-    <span class="fd-token__text">
-        Dolor
-    </span>
-</span>
-<span class="fd-token fd-token--readonly" role="button" tabindex="0">
-    <span class="fd-token__text">
-        Filter
-    </span>
-</span>
