--- packages/styles/stories/Components/token/selected.example.html
@@ -1,24 +1,0 @@
-<span class="fd-token fd-token--selected" role="button" tabindex="0">
-    <span class="fd-token__text">
-        Bibendum
-    </span>
-    <button class="fd-token__close" tabindex="-1" aria-label="Bibendum token close"></button>
-</span>
-<span class="fd-token fd-token--selected" role="button" tabindex="0">
-    <span class="fd-token__text">
-        Lorem
-    </span>
-    <button class="fd-token__close" tabindex="-1" aria-label="Lorem token close"></button>
-</span>
-<span class="fd-token fd-token--selected" role="button" tabindex="0">
-    <span class="fd-token__text">
-        Dolor
-    </span>
-    <button class="fd-token__close" tabindex="-1" aria-label="Dolor token close"></button>
-</span>
-<span class="fd-token fd-token--selected" role="button" tabindex="0">
-    <span class="fd-token__text">
-        Filter
-    </span>
-    <button class="fd-token__close" tabindex="-1" aria-label="Filter token close"></button>
-</span>
