+++ packages/styles/stories/Components/tokenizer/cozy-tokenizer.example.html
@@ -1,5 +1,11 @@
 <div class="fd-tokenizer ">
     <div class="fd-tokenizer__inner">
+        <input class="fd-input fd-tokenizer__input" aria-label="Tokenizer input" placeholder="Type Ahead"/>
+    </div>
+</div>
+<br/><br/>
+<div class="fd-tokenizer ">
+    <div class="fd-tokenizer__inner">
         <span class="fd-token" role="button" tabindex="0">
             <span class="fd-token__text">
                 Bibendum
@@ -29,7 +35,7 @@
                 Lorem
             </span>
         </span>
-        <input class="fd-input fd-tokenizer__input" aria-label="Tokenizer input"/>
+        <input class="fd-input fd-tokenizer__input" aria-label="Tokenizer input" placeholder="Type Ahead"/>
     </div>
 </div>
 <br/><br/>
